Isolate config logic and add e2e config tests
This commit puts the serialised version of the config file into one
single Rust file. This file is put under a CODEOWNERS rule so that it is
carefully check for breaking changes.

It also adds one end-to-end config test checking the behaviour of
various fields.
