+++ .github/CODEOWNERS
@@ -1,0 +1,3 @@
+# The content of the config.rs file should not change in a breaking way.
+# See https://github.com/parallaxsecond/parsec/issues/393 for details.
+src/utils/config.rs   @parallaxsecond/admin
