+++ src/utils/mod.rs
@@ -1,8 +1,9 @@
 // Copyright 2019 Contributors to the Parsec project.
 // SPDX-License-Identifier: Apache-2.0
 //! Service utilities
+pub mod config;
 mod global_config;
 mod service_builder;
 
 pub use global_config::GlobalConfig;
-pub use service_builder::{CoreSettings, ServiceBuilder, ServiceConfig};
+pub use service_builder::ServiceBuilder;
